feat: created update route

+++ app/routers/issues.py
@@ -1,4 +1,4 @@
-from fastapi import APIRouter, Request, Depends
+from fastapi import APIRouter, Request, Depends, HTTPException
 from fastapi.responses import HTMLResponse
 from schemas.issue import *
 from services.issues import *
@@ -13,7 +13,7 @@ async def postIssue(issue: CreateIssue, db: Session = Depends(getDB) ):
 
 
 @router.get('/{user_id}')
-async def getIssuesByUser(user_id: int, db: Session = Depends(getDB)):
+async def getByUser(user_id: int, db: Session = Depends(getDB)):
     return getIssuesByUser(db, user_id)
 
 @router.get('/')
@@ -21,11 +21,17 @@ async def getIssues(db: Session = Depends(getDB)):
     return  getAllIssues(db)
 
 
-@router.patch('/') 
-def patchIssue(issue: UpdateIssue):
-    context = {'request': issue}
-    return  {"message": "issues"}
+@router.patch('/{id}') 
+async def patchIssue(id: int, issue: UpdateIssue, db: Session = Depends(getDB)):
+    if not checkIfExists(db, id):
+        raise HTTPException(status_code=404, detail="ID of issue not found")
+    formattedIssue = issue.dict()
+    formattedIssue['type'] = issue.type.value
+    return updateIssue(db, id, formattedIssue)
 
 @router.delete('/{id}')
 async def delete(id: int, db: Session = Depends(getDB)):
+    if not checkIfExists(db, id):
+        raise HTTPException(status_code=404, detail="ID of issue not found")
+
     deleteIssue(db, id)
